Refactored testutils to move event helpers to a separate package
